% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dingo.R
\name{calculateDiff}
\alias{calculateDiff}
\title{table of classification performances}
\usage{
calculateDiff(n, tY.org, z, levels.z, P, w.upper)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

