% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_cibersort.R
\name{CoreAlg}
\alias{CoreAlg}
\title{CIBERSORT R script v1.03 (last updated 07-10-2015)
Note: Signature matrix construction is not currently available; use java version for full functionality.
Author: Aaron M. Newman, Stanford University (amnewman@stanford.edu)
Requirements:
      R v3.0 or later. (dependencies below might not work properly with earlier versions)
      install.packages('e1071')
      install.pacakges('parallel')
      install.packages('preprocessCore')
      if preprocessCore is not available in the repositories you have selected, run the following:
          source("http://bioconductor.org/biocLite.R")
          biocLite("preprocessCore")
Windows users using the R GUI may need to Run as Administrator to install or update packages.
This script uses 3 parallel processes.  Since Windows does not support forking, this script will run
single-threaded in Windows.}
\usage{
CoreAlg(X, y)
}
\arguments{
\item{X}{cell-specific gene expression}

\item{y}{mixed expression per sample}
}
\description{
Usage:
      Navigate to directory containing R script
}
\details{
In R:
      source('CIBERSORT.R')
      results <- CIBERSORT('sig_matrix_file.txt','mixture_file.txt', perm, QN)

Options:
      i)  perm = No. permutations; set to >=100 to calculate p-values (default = 0)
      ii) QN = Quantile normalization of input mixture (default = TRUE)

Input: signature matrix and mixture file, formatted as specified at http://cibersort.stanford.edu/tutorial.php
Output: matrix object containing all results and tabular data written to disk 'CIBERSORT-Results.txt'
License: http://cibersort.stanford.edu/CIBERSORT_License.txt
Core algorithm
}
