% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationPerf.R
\name{createFolds}
\alias{createFolds}
\title{createFolds function copied from the caret R-package}
\usage{
createFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{Y vector of phenotypic lables}

\item{k}{number of iterations of the cross-valiation}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}
