% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{enet}
\alias{enet}
\title{Elastic net classification panel}
\usage{
enet(X, Y, alpha, lambda = NULL, family, X.test = NULL, Y.test = NULL,
  filter = "p.value", topranked = 50, keepVar = NULL)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{alpha}{= 1 (lasso), alpha = 0 (ridge), 0 < alpha < 1 (elastic net penalty)}

\item{lambda}{= strength of elastic net penalty}

\item{family}{"binomial" or "multinomial"}

\item{X.test}{nxp matrx  - test dataset}

\item{filter}{= "none" or "p.value"}

\item{topranked}{= 50 (top number of features to select and build a classifier)}
}
\description{
build an elastic net classification panel
}
