% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{enetCV}
\alias{enetCV}
\title{cross-validation function for elastic net panel}
\usage{
enetCV(X, Y, alpha, lambda, M, folds, progressBar, family, filter, topranked)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{alpha}{= 1 (lasso), alpha = 0 (ridge), 0 < alpha < 1 (elastic net penalty)}

\item{lambda}{= strength of elastic net penalty}

\item{M}{= # of folds}

\item{folds}{list of length iter with each element a list of length M}

\item{family}{"binomial" or "multinomial"}

\item{filter}{= "none" or "p.value"}

\item{topranked}{= 50 (top number of features to select and build a classifier)}

\item{X.test}{nxp matrx  - test dataset}
}
\description{
Estimate test error of elastic net panel
}

