% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensemble.splsda.R
\name{ensemble.splsda}
\alias{ensemble.splsda}
\title{Build ensemble sPLSDA classification panel}
\usage{
ensemble.splsda(X.train, Y.train, keepXList, ncomp, X.test, Y.test, filter,
  topranked)
}
\arguments{
\item{X.train}{- list of training datasets (nxpi); i number of elements}

\item{Y.train}{- n-vector of class labels}

\item{ncomp}{= Number of components}

\item{X.test}{- list of test datasets (nxpi); i number of elements}

\item{Y.test}{- n-vector of class labels}

\item{keepList}{= list of keepX}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

