% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensemble.splsda.R
\name{ensemble.splsdaCV}
\alias{ensemble.splsdaCV}
\title{Estimate cross-validation error using cross-validation}
\usage{
ensemble.splsdaCV(X, Y, keepXList, ncomp, M, folds, progressBar, filter,
  topranked)
}
\arguments{
\item{X}{- list of training datasets (nxpi); i number of elements}

\item{Y}{- n-vector of class labels}

\item{keepXList}{= list of keepX}

\item{M}{- # of folds}

\item{folds}{- list of length M, where each element contains the indices for samples for a given fold}

\item{progressBar}{(TRUE/FALSE) - show progress bar or not}

\item{filter}{- "none" or "p.value"}

\item{topranked}{- # of significant features to use to build classifier}
}
\description{
Estimate cross-validation error using cross-validation
}

