% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleEnet.R
\name{ensembleCV}
\alias{ensembleCV}
\title{Estimate cross-validation using cross-validation}
\usage{
ensembleCV(X.trainList, Y.train, alphaList, lambdaList, family, filter,
  topranked, keepVarList, M, folds, progressBar)
}
\arguments{
\item{filter}{- "none" or "p.value"}

\item{topranked}{- # of significant features to use to build classifier}

\item{M}{- # of folds}

\item{folds}{- list of length M, where each element contains the indices for samples for a given fold}

\item{progressBar}{(TRUE/FALSE) - show progress bar or not}

\item{X}{- list of training datasets (nxpi); i number of elements}

\item{Y}{- n-vector of class labels}

\item{alpha}{= list of alpha values}

\item{lambda}{= list of lambda values}
}
\description{
Estimate cross-validation using cross-validation
}
