% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_diablo.R
\name{extractErr}
\alias{extractErr}
\title{clean up diablo perf() output}
\usage{
extractErr(cv)
}
\arguments{
\item{cv}{- perf(obj) - obj-block.splsda object}
}
\description{
clean up diablo perf() output
}
