% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dingo.R
\name{fast.dingo}
\alias{fast.dingo}
\title{table of classification performances}
\usage{
fast.dingo(dat, x, rhoarray = NULL, diff.score = T, B = 30, verbose = T,
  threads = 4)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

