% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dingo.R
\name{fast.scoring.boot}
\alias{fast.scoring.boot}
\title{recompute differential edge scores}
\usage{
fast.scoring.boot(stddat, z, Omega, A, B, boot.B = 100, verbose = T,
  threads)
}
\description{
recompute differential edge scores
}

