% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_analysis.R
\name{gene_set_analysis}
\alias{gene_set_analysis}
\title{Gene Set Analysis (GSA)}
\usage{
gene_set_analysis(genes, genesets, study_universe, min.set.size = 1)
}
\arguments{
\item{genesets}{list of gene sets}

\item{genelist}{vector of gene symbols}

\item{universe}{background list of genes: all the genes used in the study}
}
\value{
The output is a dataframe with 4 columns (gene sets (genesets), p-value (pval), bhfdr (BH-FDR), Type (unfiltered or filtered)
}
\description{
\code{gene_set_analysis} Perform gene set analysis using the hypergeometric test
}
\details{
This function uses the hypergeometric function to test for over-representation of genesets in a user-specified list of genes. The unfiltered genesets contain all genes that are part of certain databases, whereas filtered genesets only contain genes that were used in the experiment (in order prevent sampling bias, \url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-015-0761-7}).
}
\examples{

library(dplyr)
data(pathways)
# gene list supplied by the user
set.seed(1)
genelist <- pathways \%>\%
 filter(Database == "KEGG") \%>\%
 dplyr::select(Genes) \%>\%
 unlist \%>\%
 as.character \%>\%
 sample(., 50)

set.seed(2)
universe <- pathways \%>\%
 filter(Database == "KEGG") \%>\%
 dplyr::select(Genes) \%>\%
 unlist \%>\%
 as.character \%>\%
 sample(., 50) \%>\%
 c(., genelist) \%>\%
 unique

# Perform gene set enrichment analysis
result <- gene_set_analysis(genes = genelist,
 genesets = pathways,
 study_universe = universe,
 min.set.size = 1)

result$toptable \%>\% group_by(Type, Database) \%>\% slice(1:5)

}
