% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_glm.R
\name{glmCV}
\alias{glmCV}
\title{interal function (glm cross-validation)}
\usage{
glmCV(X, Y, M, folds, progressBar)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{M}{= # of folds}

\item{folds}{list of length M specifiying which samples are in which fold}

\item{progressBar}{(TRUE/FALSE)}
}
\description{
Estimate test error of glm biomarker
}

