% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_glm.R
\name{glmPanel}
\alias{glmPanel}
\title{GLM single biomarker}
\usage{
glmPanel(X, Y, X.test = NULL, Y.test = NULL)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{X.test}{nxp matrx  - test dataset}

\item{Y.test}{lablel of test dataset}
}
\description{
build a glm biomarker
}
