% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkFunctions.R
\name{graphIndices}
\alias{graphIndices}
\title{graphIndices()}
\usage{
graphIndices(panels = panels, X.train = X.train, cutoff = cutoff)
}
\arguments{
\item{panels}{- list of panels (character vector corresponding to column names of X.train)}

\item{X.train}{- list of datasets (n x p)}

\item{cut-off}{- of person correlatio}

\item{concat_lambda}{- value that controls the strength of the penalization}

\item{single_alphaList}{- list of alpha values of length k}

\item{single_lambdaList}{- list of lambda values of length k}
}
\description{
build integrative panels (Concatenation, Ensemble, DIABLO)
}
