% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_analysis.R
\name{hypertest}
\alias{hypertest}
\title{hypergeometric Test}
\usage{
hypertest(genelist, genesets, cpus)
}
\arguments{
\item{genelist}{vector of gene symbols}

\item{genesets}{list of gene sets}
}
\value{
The output is a dataframe with 3 columns (gene sets (genesets), p-value (pval), and bhfdr (BH-FDR)
}
\description{
\code{hypertest} performs hypergeometric tests across a list of gene sets
}
\details{
This function uses the hypergeometric function to test for over-representation of genesets in a user-specified list of genes.
}
