% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkFunctions.R
\name{integrativePanels}
\alias{integrativePanels}
\title{integrativePanels()}
\usage{
integrativePanels(X.train, Y.train, X.test, Y.test, concat_alpha, concat_lambda,
  single_alphaList, single_lambdaList, M, iter, threads)
}
\arguments{
\item{X.train}{list of nxp datasets of length k}

\item{Y.train}{- n-vector of class labels}

\item{concat_alpha}{- 0 < value < 1}

\item{concat_lambda}{- value that controls the strength of the penalization}

\item{single_alphaList}{- list of alpha values of length k}

\item{single_lambdaList}{- list of lambda values of length k}
}
\description{
build integrative panels (Concatenation, Ensemble, DIABLO)
}
