% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkFunctions.R
\name{networkStats}
\alias{networkStats}
\title{networkStats()}
\usage{
networkStats(adjMat, mode = "graph")
}
\arguments{
\item{adjMat}{p x p matrix}

\item{mode}{= "graph"}
}
\description{
compuate network statistics given a adjacency matrix
}

