% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda_functions.R
\name{pandaModif}
\alias{pandaModif}
\title{run PANDA (calculate geneCoreg, regNet and co-operative matrices)}
\usage{
pandaModif(motif, expr = NULL, ppi = NULL, alpha = 0.1,
  hamming = 0.01, iter = NA)
}
\arguments{
\item{motif}{3 column matrix, 1st column (regVar), 2nd column (gene), 3rd (weight)}

\item{geneExp}{nxp matrix (gene by sample)}

\item{regulatory}{expression (regVar by sample)}

\item{phenotype}{factor specifiying the group of each sample}
}
\description{
run PANDA (calculate geneCoreg, regNet and co-operative matrices)
}
