% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda_functions.R
\name{pandax}
\alias{pandax}
\title{Caculate difference between regulatory coefficients of the two groups}
\usage{
pandax(motif, geneExp, regExp, phenotype)
}
\arguments{
\item{motif}{3 column matrix, 1st column (regVar), 2nd column (gene), 3rd (weight)}

\item{geneExp}{nxp matrix (gene by sample)}

\item{phenotype}{factor specifiying the group of each sample}

\item{regulatory}{expression (regVar by sample)}
}
\description{
Caculate difference between regulatory coefficients of the two groups
}

