% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda_functions.R
\name{pandaxBoot}
\alias{pandaxBoot}
\title{Caculate differential regulatory network statistics}
\usage{
pandaxBoot(motif, geneExp, regExp, phenotype, nperms)
}
\arguments{
\item{motif}{3 column matrix, 1st column (regVar), 2nd column (gene), 3rd (weight)}

\item{geneExp}{nxp matrix (gene by sample)}

\item{phenotype}{factor specifiying the group of each sample}

\item{nperms}{# of permutations}

\item{regulatory}{expression (regVar by sample)}
}
\description{
Caculate differential regulatory network statistics
}

