% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{table of classification performances}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}
