% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{perf.enet}
\alias{perf.enet}
\title{table of classification performances}
\usage{
perf.enet(object, validation = c("Mfold", "loo"), M = 5, iter = 10,
  threads = 4, progressBar = TRUE)
}
\arguments{
\item{object}{- elastic net classifier}

\item{validation}{- Mfold or LOOCV}

\item{M}{- # of folds}

\item{iter}{- number of times to repeat cross-validation}

\item{threads}{- number of cpus (each cross-valiation scheme performed on a separate node)}

\item{progressBar}{- show progressbar (TRUE/FALE)}
}
\description{
build an elastic net classification panel
}

