% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleEnet.R
\name{perf.ensembleEnet}
\alias{perf.ensembleEnet}
\title{Estimate classification performance using repeated cross-validation}
\usage{
perf.ensembleEnet(object, validation = "Mfold", M = 5, iter = 5,
  threads = 5, progressBar = TRUE)
}
\arguments{
\item{object}{- ensembleEnet object}

\item{validation}{= "Mfold" or "loo"}

\item{M}{- # of folds}

\item{iter}{- Number of iterations of cross-validation}

\item{threads}{- # of cores, running each iteration on a separate node}

\item{progressBar}{= TRUE (show progress bar or not)}
}
\description{
Estimate classification performance using repeated cross-validation
}
