% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleRf.R
\name{perf.ensembleRf}
\alias{perf.ensembleRf}
\title{Estimate classification performance using repeated cross-validation using an random forest classifier}
\usage{
perf.ensembleRf(object, validation = "Mfold", M = 5, iter = 5,
  threads = 5, progressBar = TRUE)
}
\arguments{
\item{object}{ensembleRf object}

\item{validation}{= "Mfold" or "loo"}

\item{M}{# of folds}

\item{iter}{Number of iterations of cross-validation}

\item{threads}{# of cores, running each iteration on a separate node}

\item{progressBar}{= TRUE (show progress bar or not)}
}
\description{
Estimate classification performance using repeated cross-validation using an random forest classifier
}
