% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_glm.R
\name{perf.glm}
\alias{perf.glm}
\title{cross-validation function for glm}
\usage{
perf.glm(object, validation = c("Mfold", "loo"), M = 5, iter = 10,
  threads = 4, progressBar = TRUE)
}
\arguments{
\item{object}{= glmPanel object}

\item{validation}{- Mfold or LOOCV}

\item{M}{= # of folds}

\item{iter}{= number of times to repeat cross-valiation}

\item{threads}{- number of cpus (each cross-valiation scheme performed on a separate node)}

\item{progressBar}{- show progressbar (TRUE/FALE)}
}
\description{
Estimate test error of glm biomarker
}

