% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_rf.R
\name{perf.rf}
\alias{perf.rf}
\title{Cross-validation repeated iter number of times}
\usage{
perf.rf(object, validation = c("Mfold", "loo"), M = 5, iter = 10,
  threads = 4, progressBar = TRUE)
}
\arguments{
\item{object}{= rforest object}

\item{validation}{= "Mfold" or "loo}

\item{M}{= # of folds}

\item{iter}{= number of times to repeat cross-validation}

\item{threads}{= number of cores to use (each cross-valiation scheme performed on a separate node)}

\item{show}{progress bar or not: progressBar (TRUE/FALSE)}
}
\description{
Estimate the mean+/-SD test error
}
