% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{perf.splsda}
\alias{perf.splsda}
\title{repeated CV function for a splsda model}
\usage{
perf.splsda(object, validation = c("Mfold", "loo"), M = 5, iter = 10,
  threads = 4, progressBar = TRUE)
}
\arguments{
\item{object}{splsda model object}

\item{M}{Number of folds in the cross-validation}

\item{iter}{number of times to repeat the cross-validation}

\item{threads}{number of cores to use to parrallel the CV}

\item{progressBar}{display progress bar or not (TRUE/FALSE)}

\item{valdation}{type of cross-validation; Mfold "Mfold" or LOOCV "loo"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

