% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{perf.splsda2}
\alias{perf.splsda2}
\title{table of classification performances}
\usage{
perf.splsda2(object, method.predict = c("all", "max.dist", "centroids.dist",
  "mahalanobis.dist"), validation = c("Mfold", "loo"), folds = 10,
  progressBar = TRUE, near.zero.var = FALSE)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

