% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensemble.splsda.R
\name{perfEnsemble.splsda}
\alias{perfEnsemble.splsda}
\title{Estimate test error using repeated cross-validation}
\usage{
perfEnsemble.splsda(object, validation = "Mfold", M = M, iter = iter,
  threads = threads, progressBar = TRUE)
}
\arguments{
\item{object}{- ensembleEnet object}

\item{validation}{= Mfold or loo}

\item{M}{- # of folds}

\item{iter}{- Number of iterations of cross-validation}

\item{threads}{- # of cores, running each iteration on a separate node}

\item{progressBar}{= TRUE (show progress bar or not)}
}
\description{
Estimate test error using repeated cross-validation
}

