% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_nanoString.R
\name{rccToDat}
\alias{rccToDat}
\title{Convert .RCC file to data frame}
\usage{
rccToDat(fileName)
}
\arguments{
\item{fileName}{file name including file path}
}
\description{
NanoString .RCC files to data frame
}
