% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_rf.R
\name{rforest}
\alias{rforest}
\title{Random forest classification panel}
\usage{
rforest(X, Y, X.test = NULL, Y.test = NULL, family = "binomial",
  filter = "p.value", topranked = 50)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{X.test}{nxp matrx  - test dataset}

\item{family}{"binomial" or "multinomial"}

\item{filter}{= "none" or "p.value"}

\item{topranked}{= 50 (top number of features to select and build a classifier)}
}
\description{
build random forest classification panel
}

