% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_diablo.R
\name{setDesign}
\alias{setDesign}
\title{clean up diablo perf() output}
\usage{
setDesign(X, corCutOff, plotMat = FALSE)
}
\arguments{
\item{cv}{- perf(obj) - obj-block.splsda object}
}
\description{
clean up diablo perf() output
}

