% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{splitData}
\alias{splitData}
\title{split demographics into two datasets (continuous and categorical variables)}
\usage{
splitData(demo, group, trim = 0.8)
}
\arguments{
\item{demo}{demographics dataset}

\item{group}{string stateting the column name of the grouping variable}

\item{trim}{remove variables with 20% missing data}
}
\description{
split demographics into two datasets (continuous and categorical variables)
}
