% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{sPLSDA}
\alias{sPLSDA}
\title{sPLSDA function}
\usage{
sPLSDA(X.train, Y.train, keepX, ncomp, X.test = NULL, Y.test = NULL,
  filter = "p.value", topranked = 50)
}
\arguments{
\item{X.train}{n1xp dataset (discovery cohort)}

\item{Y.train}{vector of phenotype labels with names(Y.train) == rownames(X.train)}

\item{keepX}{(# of variables to select per component)}

\item{ncomp}{number of components}

\item{X.test}{n2xp dataset (validaton cohort)}

\item{Y.test}{vector of phenotype labels with names(Y.test) == rownames(X.test)}

\item{filter}{apply no filter "none" or a p.value filter "p.value}

\item{topranked}{select the top significant variables based on limma}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}
