% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{splsdaCV}
\alias{splsdaCV}
\title{CV function for a splsda model}
\usage{
splsdaCV(X, Y, keepX, ncomp, M, folds, progressBar, filter, topranked)
}
\arguments{
\item{X}{nxp dataset}

\item{Y}{vector of phenotype labels with names(Y) == rownames(X)}

\item{keepX}{(# of variables to select per component)}

\item{ncomp}{number of components}

\item{M}{Number of folds in the cross-validation}

\item{folds}{list of elements (sample indices) for the M folds}

\item{progressBar}{display progress bar or not (TRUE/FALSE)}

\item{filter}{apply no filter "none" or a p.value filter "p.value}

\item{topranked}{select the top significant variables based on limma}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

