% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_svm.R
\name{supportVectorMachine}
\alias{supportVectorMachine}
\title{SVM classification panel}
\usage{
supportVectorMachine(X, Y, X.test = NULL, Y.test = NULL,
  family = "binomial", filter = "p.value", topranked = 50)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{X.test}{nxp matrx  - test dataset}

\item{family}{"binomial" or "multinomial"}

\item{filter}{= "none" or "p.value"}

\item{topranked}{= 50 (top number of features to select and build a classifier)}

\item{alpha}{= 1 (lasso), alpha = 0 (ridge), 0 < alpha < 1 (elastic net penalty)}

\item{lambda}{= strength of elastic net penalty}
}
\description{
build a SVM classification panel
}

