% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationPerf.R
\name{tperformance}
\alias{tperformance}
\title{table of classification performances}
\usage{
tperformance(weights, trueLabels, direction)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{direction}{= "auto", ">", "<"}

\item{trubeLabels}{are the true labels associated with the test data}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

