% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{tuned.spslda}
\alias{tuned.spslda}
\title{splsda model after tuning of the number of variables}
\usage{
tuned.spslda(X, Y, ncomp, keepXgrid, validation, M)
}
\arguments{
\item{X}{nxp dataset}

\item{Y}{vector of phenotype labels with names(Y) == rownames(X)}

\item{ncomp}{number of components}

\item{keepXgrid}{sequence of integers (# of variables to select per component)}

\item{M}{Number of folds in the cross-validation}

\item{validatoin}{"Mfold" or "loo"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

