% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{tuned.spsldaCV}
\alias{tuned.spsldaCV}
\title{Runs a cross-validation scheme for a tuned.splsda model once}
\usage{
tuned.spsldaCV(X = X, Y = Y, ncomp = ncomp, folds = folds,
  progressBar = FALSE, keepXgrid = keepXgrid, validation = validation,
  M = M)
}
\arguments{
\item{X}{nxp dataset}

\item{Y}{vector of phenotype labels with names(Y) == rownames(X)}

\item{ncomp}{number of components}

\item{folds}{list of length M where each element contains the indices of the samples in that fold}

\item{progressBar}{= TRUE/FALSE}

\item{keepXgrid}{sequence of integers (# of variables to select per component)}

\item{M}{Number of folds in the cross-validation}

\item{validatoin}{"Mfold" or "loo"}
}
\description{
estimate the error rate and AUC of tuned.splsda panel
}

