% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{venndiagram}
\alias{venndiagram}
\title{Venn Diagram}
\usage{
venndiagram(datList, circleNames)
}
\arguments{
\item{datList}{list of character vectors}

\item{circleNames}{names of list elements}
}
\description{
Venn diagram of 2, 3, or 4 sets
}

