% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/supportFunc_ggplot2.R
\name{customTheme}
\alias{customTheme}
\title{table of classification performances}
\usage{
customTheme(sizeStripFont, xAngle, hjust, vjust, xSize, ySize, xAxisSize,
  yAxisSize)

customTheme(sizeStripFont, xAngle, hjust, vjust, xSize, ySize, xAxisSize,
  yAxisSize)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}

\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs

takes in predited weights and true labels and determines performance characterisitcs
}

