% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{enet}
\alias{enet}
\title{table of classification performances}
\usage{
enet(X, Y, alpha, lambda = NULL, family, X.test = NULL, Y.test = NULL)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

