% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{enetCV}
\alias{enetCV}
\title{interal function (enet cross-validation)}
\usage{
enetCV(X, Y, alpha, lambda, M, folds, progressBar, family, filter, topranked)
}
\arguments{
\item{M}{- # of folds}

\item{progressBar}{- show progressbar (TRUE/FALE)}

\item{object}{- elastic net classifier}

\item{validation}{- Mfold or LOOCV}

\item{iter}{- number of times to repeat cross-validation}

\item{threads}{- number of cpus (each cross-valiation scheme performed on a separate node)}
}
\description{
build an elastic net classification panel
}

