% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{enetCV}
\alias{enetCV}
\title{interal function (enet cross-validation)}
\usage{
enetCV(X, Y, alpha, lambda, M, folds, progressBar, family, filter, topranked,
  keepVar)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{M}{= # of folds}

\item{folds}{list of length M specifiying which samples are in which fold}

\item{family}{"binomial" or "multinomial"}

\item{filter}{= "none" or "p.value"}

\item{topranked}{= 50 (top number of features to select and build a classifier)}
}
\description{
Estimate test error of elastic net panel
}
