% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleEnet.R
\name{ensembleCV}
\alias{ensembleCV}
\title{Estimate cross-validation using cross-validation}
\usage{
ensembleCV(X, Y, alpha, lambda, M, folds, progressBar, filter, topranked)
}
\arguments{
\item{X}{- list of training datasets (nxpi); i number of elements}

\item{Y}{- n-vector of class labels}

\item{alpha}{= list of alpha values}

\item{lambda}{= list of lambda values}

\item{M}{- # of folds}

\item{folds}{- list of length M, where each element contains the indices for samples for a given fold}

\item{progressBar}{(TRUE/FALSE) - show progress bar or not}

\item{filter}{- "none" or "p.value"}

\item{topranked}{- # of significant features to use to build classifier}
}
\description{
Estimate cross-validation using cross-validation
}

