% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleEnet.R
\name{ensembleEnet}
\alias{ensembleEnet}
\title{Build ensemble classification panel}
\usage{
ensembleEnet(X.train, Y.train, alphaList, lambdaList, X.test, Y.test, filter,
  topranked)
}
\arguments{
\item{X.train}{- list of training datasets (nxpi); i number of elements}

\item{Y.train}{- n-vector of class labels}

\item{X.test}{- list of test datasets (nxpi); i number of elements}

\item{Y.test}{- n-vector of class labels}

\item{alpha}{= list of alpha values}

\item{lambda}{= list of lambda values}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}
