% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleEnet.R
\name{ensembleEnet}
\alias{ensembleEnet}
\title{Build ensemble enet classification panel}
\usage{
ensembleEnet(X.trainList, y.train, alphaList, lambdaList,
  family = "binomial", X.testList = NULL, y.test = NULL,
  filter = "none", topranked = 50, keepVarList = NULL)
}
\arguments{
\item{X.trainList}{- list of training datasets (nxpi); i number of elements}

\item{y.train}{- n-vector of class labels (must be a factor)}

\item{alphaList}{= list of alpha values}

\item{lambdaList}{= list of lambda values}

\item{family}{- can be "binomial" or "multinomial"}

\item{X.testList}{- list of test datasets (nxpi); i number of elements}

\item{y.test}{- n-vector of class labels (must be a factor)}

\item{filter}{- pre-filtering of initial datasets - "none" or "p.value"}

\item{topranked}{- Number of topranked features based on differential expression to use to build classifer}

\item{keepVarList}{- which variables to keep and not omit (set to NULL if no variables are forced to be kept)}
}
\value{
model

testPerf

X.trainList

y.train

alphaList

lambdaList

family

X.testList

y.test

filter

topranked

keepVarList
}
\description{
Build ensemble enet classification panel
}
