% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleRf.R
\name{ensembleRf}
\alias{ensembleRf}
\title{table of classification performances}
\usage{
ensembleRf(X.train, Y.train, X.test, Y.test)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}
