% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleRf.R
\name{ensembleRfCV}
\alias{ensembleRfCV}
\title{Estimate classification performance using cross-validation using an random forest classifier}
\usage{
ensembleRfCV(X.trainList, y.train, panelLength = 15, filter = "none",
  topranked = 50, keepVarList = NULL, M = 5, folds = 5,
  progressBar = FALSE)
}
\arguments{
\item{X.trainList}{list of training datasets (nxpi); i number of elements}

\item{y.train}{n-vector of class labels (must be a factor)}

\item{filter}{pre-filtering of initial datasets - "none" or "p.value"}

\item{topranked}{Number of topranked features based on differential expression to use to build classifer}

\item{keepVarList}{which variables to keep and not omit (set to NULL if no variables are forced to be kept)}

\item{M}{# of folds}

\item{folds}{list of length M, where each element contains the indices for samples for a given fold}

\item{progressBar}{(TRUE/FALSE) - show progress bar or not}

\item{alphaList}{list of alpha values}

\item{lambdaList}{list of lambda values}

\item{family}{can be "binomial" or "multinomial"}
}
\value{
error computes error rate (each group, overall and balanced error rate)

perfTest classification performance measures
}
\description{
Estimate classification performance using cross-validation using an random forest classifier
}
