% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dingo.R
\name{fast.dingo}
\alias{fast.dingo}
\title{Perform DINGO}
\usage{
fast.dingo(dat, x, rhoarray = NULL, diff.score = T, B = 30, verbose = T,
  threads = 4)
}
\arguments{
\item{dat}{nxp dataset}

\item{x}{response vector of n subjects}

\item{rhoarray}{penalization values for inverse covariance matrix (computed via glasso)}

\item{diff.score}{compute edge score}

\item{B}{Number of permutations (bootstraps)}

\item{verbose}{TRUE/FALSE}

\item{threads}{Number of cores to permutations over cores}
}
\description{
Differential Network Analysis in Genomics
}

