% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_analysis.R
\name{gene_set_analysis}
\alias{gene_set_analysis}
\title{Gene Set Analysis (GSA)}
\usage{
gene_set_analysis(genelist, genesets, universe)
}
\arguments{
\item{genelist}{vector of gene symbols}

\item{genesets}{list of gene sets}

\item{universe}{background list of genes: all the genes used in the study}
}
\value{
The output is a dataframe with 4 columns (gene sets (genesets), p-value (pval), bhfdr (BH-FDR), Type (unfiltered or filtered)
}
\description{
\code{gene_set_analysis} Perform gene set analysis using the hypergeometric test
}
\details{
This function uses the hypergeometric function to test for over-representation of genesets in a user-specified list of genes. The unfiltered genesets contain all genes that are part of certain databases, whereas filtered genesets only contain genes that were used in the experiment (in order prevent sampling bias, \url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-015-0761-7}).
}
\examples{

data(pathways)

# genelist and KEGG pathways (gene sets)
set.seed(1)
genelist <- sample(unlist(unlist(kegg)), 50)
set.seed(2)
universe <- unique(c(genelist, sample(unlist(unlist(kegg)), 300)))

# Perform gene set enrichment analysis
result <- gene_set_analysis(genelist, genesets, universe)

result \%>\% group_by(Type) \%>\% slice(1:5)
}
