% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{hypothesisTests}
\alias{hypothesisTests}
\title{perform hypothesis test based on variable type}
\usage{
hypothesisTests(data, group)
}
\arguments{
\item{data}{input dataset}

\item{group}{string stateting the column name of the grouping variable}
}
\description{
perform hypothesis test based on variable type
}
