% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_enet.R
\name{perf.enet}
\alias{perf.enet}
\title{cross-validation function for elastic net panel}
\usage{
perf.enet(object, validation = c("Mfold", "loo"), M = 5, iter = 10,
  threads = 4, progressBar = TRUE)
}
\arguments{
\item{M}{= # of folds}

\item{iter}{= number of times to repeat cross-valiation}

\item{threads}{- number of cpus (each cross-valiation scheme performed on a separate node)}

\item{progressBar}{- show progressbar (TRUE/FALE)}

\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{alpha}{= 1 (lasso), alpha = 0 (ridge), 0 < alpha < 1 (elastic net penalty)}

\item{family}{"binomial" or "multinomial"}

\item{lambda}{= strength of elastic net penalty}
}
\description{
Estimate test error of elastic net panel
}
