% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_splsda.R
\name{perf.splsda2}
\title{Prediction fuction for splsda}
\usage{
perf.splsda2(object, method.predict = c("all", "max.dist", "centroids.dist",
  "mahalanobis.dist"), validation = c("Mfold", "loo"), folds = 10,
  progressBar = TRUE, near.zero.var = FALSE)
}
\arguments{
\item{object}{splsda model}

\item{method.predict}{type of distance measure to use: "max.dist", "centroids.dist", "mahalanobis.dist" or "all"}

\item{validation}{Either an M-fold cross-valiadtion where M is an integr between 1-n (total number of samples)}

\item{folds}{Number of folds, used only if validation = "Mfold"}

\item{progressBar}{display progress bar (TRUE) or Not (FALSE)}
}
\description{
predict scores for newdata
}
\examples{
\dontrun{

}
}

