% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_svm.R
\name{perf.svm}
\alias{perf.svm}
\title{table of classification performances}
\usage{
perf.svm(object, validation = c("Mfold", "loo"), M = 5, iter = 10,
  threads = 4, progressBar = TRUE)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

