% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunc_ensembleEnet.R
\name{perfEnsemble}
\alias{perfEnsemble}
\title{table of classification performances}
\usage{
perfEnsemble(object, validation = "Mfold", M = M, iter = iter,
  threads = threads, progressBar = TRUE)
}
\arguments{
\item{weights}{are the predicted scores/probablities of test data}

\item{trubeLabels}{are the true labels associated with the test data}

\item{direction}{= "auto", ">", "<"}
}
\description{
takes in predited weights and true labels and determines performance characterisitcs
}

